# perf: switch from object spread to `Object.assign` when merging globals