@alfaproject 2022-09-25
Am I reading your numbers wrongly or the times got worse after the change?