@mdjermanovic 2022-09-26
@alfaproject in these tests, less ms is better. Linting a test project that has 450 files (the Multi Files test) takes 21488.7766ms before the change, 13452.5753ms after the change, so it works faster after the change.