@alfaproject 2022-09-28
Ah sorry, I didn't realise there was a multi files test. Only noticed the single ones, never mind then